refactor: (scripts): modularizing the index.js endpoints.
